defined MinGfx specific trig functions
To avoid complications encountered with different compilers and
language standards, the float returning trigonometric functions
have been added to the MinGfx::GfxMath class to contain their
intended behavior. Compiler directives are used to switch between
which version of the function to call based on the platform or
compiler that is being used.

changed trig functions to MinGfx specific ones

The original problem was rooted in the quickshapes file, so this
is on a separate commit.

changed trig functions to MinGfx specific ones

added imports for new trig function definitions